requirements.txt
labelImg
ultralytics==8.1.26
flask
opencv-python
numpy
Pillow